Project: Make the login form not look like a pile of crap
Tailwind is a plague on front-end development. To do this correctly the Tailwind way would mean reproducing every single one of allauth's templates, and replacing `<h1>` with `<h1 class="text-4xl text-gray-700 font-black>`. What sort of insanity does one have to suffer to think that having to add that many presentational classes to recreate the basic functionality that every web browser gives you out of the box is a good idea? So, rather than the correct approach, which would be to rip out all of the bloated Tailwind markup, replace it with solid, semantic markup, and write a decent stylesheet for it all, I just added a `<style>` tag to the base template that tries to fix some of Tailwind's most egregious sins.

style

blech